tratamento de erro nos deletes e corrigindo erros de recursividade

+++ biblioteca/src/main/java/com/residencia/biblioteca/services/AlunoService.java
@@ -33,7 +33,23 @@ public class AlunoService {
 	}
 	
 	//deletar um determinado aluno
-	public void deletarAluno(Aluno aluno) {
+	public Boolean deletarAluno(Aluno aluno) {
+		if(aluno == null)
+			return false;
+		
+		Aluno alunoExistente = buscarAlunoPorId(aluno.getNumeroMatriculaAluno());
+		
+		if(alunoExistente == null)
+			return false;
+		
 		alunoRepo.delete(aluno);
+		
+		Aluno alunoContinuaExistindo =
+				buscarAlunoPorId(aluno.getNumeroMatriculaAluno());
+		
+		if(alunoContinuaExistindo != null)
+			return false;
+		
+		return true;
 	}
 }
